back-end/requirements.txt
opencv-python
mediapipe==0.10.9
numpy
pyttsx3
flask
flask-cors
flask-socketio
flask-limiter
flask-talisman
flask-wtf
scikit-learn
pymongo
python-dotenv
bcrypt
pyjwt
eventlet
email-validator
redis
itsdangerous